+++ index.html
@@ -13,6 +13,7 @@
             rel="stylesheet"
         />
         <link rel="stylesheet" href="styles/difficultyLevel.css" />
+        <link rel="stylesheet" href="game_style.css"/>
         <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Poppins" />
     </head>
     <body>
